bump dependencies, update lint rules, de-nest component
