+++ .oxlintrc.json
@@ -106,6 +106,7 @@
       "error",
       { "checkFragmentShorthand": true, "checkKeyMustBeforeSpread": true, "warnOnDuplicates": true }
     ],
+    "react/jsx-max-depth": ["error", { "max": 8 }],
     "react/jsx-no-duplicate-props": "error",
     "react/jsx-no-undef": "error",
     "react/jsx-no-useless-fragment": "error",
@@ -179,14 +180,9 @@
     "react-doctor/jsx-fragments": "error",
     "react-doctor/jsx-handler-names": "error",
     "react-doctor/jsx-key": "error",
-    "react-doctor/jsx-max-depth": "error",
     "react-doctor/jsx-no-comment-textnodes": "error",
     "react-doctor/jsx-no-constructed-context-values": "error",
     "react-doctor/jsx-no-duplicate-props": "error",
-    "react-doctor/jsx-no-jsx-as-prop": "error",
-    "react-doctor/jsx-no-new-array-as-prop": "error",
-    "react-doctor/jsx-no-new-function-as-prop": "error",
-    "react-doctor/jsx-no-new-object-as-prop": "error",
     "react-doctor/jsx-no-script-url": "error",
     "react-doctor/jsx-no-undef": "error",
     "react-doctor/jsx-no-useless-fragment": "error",
@@ -280,7 +276,6 @@
     "react-doctor/no-long-transition-duration": "error",
     "react-doctor/no-mirror-prop-effect": "error",
     "react-doctor/no-moment": "error",
-    "react-doctor/no-multi-comp": "error",
     "react-doctor/no-mutable-in-deps": "error",
     "react-doctor/no-mutating-reducer-state": "error",
     "react-doctor/no-namespace": "error",
@@ -320,7 +315,6 @@
     "react-doctor/no-unescaped-entities": "error",
     "react-doctor/no-unknown-property": "error",
     "react-doctor/no-unsafe": "error",
-    "react-doctor/no-unstable-nested-components": "error",
     "react-doctor/no-usememo-simple-expression": "error",
     "react-doctor/no-wide-letter-spacing": "error",
     "react-doctor/no-will-update-set-state": "error",
