--- components/Package/CommunitySection/MorePackagesBox.tsx
@@ -1,139 +1,0 @@
-import { LI, UL } from '@expo/html-elements';
-import { startCase } from 'es-toolkit/string';
-import { type NextPageContext } from 'next';
-import * as emoji from 'node-emoji';
-import { View } from 'react-native';
-import useSWR from 'swr';
-
-import { A, Caption, H6Section, Label, useLayout } from '~/common/styleguide';
-import { DownloadIcon, StarIcon, WarningIcon } from '~/components/Icons';
-import EntityCounter from '~/components/Package/EntityCounter';
-import ThreeDotsLoader from '~/components/Package/ThreeDotsLoader';
-import { Tooltip } from '~/components/Tooltip';
-import { type APIResponseType, type LibraryType } from '~/types';
-import { TimeRange } from '~/util/datetime';
-import getApiUrl from '~/util/getApiUrl';
-import { bigNumberFormatter, NUMBER_FORMATTER } from '~/util/strings';
-import tw from '~/util/tailwind';
-import urlWithQuery from '~/util/urlWithQuery';
-
-type Props = {
-  library: LibraryType;
-};
-
-const LIMIT = 6;
-
-export default function MorePackagesBox({ library }: Props) {
-  const owner = library.github.fullName.split('/')[0];
-
-  const { isSmallScreen } = useLayout();
-  const { data, isLoading } = useSWR<APIResponseType>(
-    getApiUrl(
-      urlWithQuery(`/libraries`, { owner, order: 'downloads', limit: LIMIT.toString() }),
-      {} as NextPageContext
-    ),
-    (url: string) =>
-      fetch(url).then(res => {
-        if (res.status === 200) {
-          return res.json();
-        }
-        return { libraries: [], total: 0 };
-      }),
-    {
-      dedupingInterval: TimeRange.HOUR * 1000,
-      revalidateOnFocus: false,
-    }
-  );
-
-  if (data && data.libraries.length <= 1) {
-    return null;
-  }
-
-  return (
-    <>
-      <H6Section style={[tw`flex items-center gap-1.5`, !isSmallScreen && tw`mt-4`]}>
-        More packages from {startCase(owner)}
-        {!isLoading && data?.total && data.total > 0 ? (
-          <EntityCounter count={data.total > LIMIT ? data.total : data.total - 1} />
-        ) : null}
-        {!isSmallScreen && data?.total && data.total > LIMIT && (
-          <A href={`/packages?owner=${encodeURI(owner)}`} style={tw`ml-auto`}>
-            <Label style={tw`font-light`}>See all packages</Label>
-          </A>
-        )}
-      </H6Section>
-      {!data || isLoading ? (
-        <ThreeDotsLoader />
-      ) : (
-        <>
-          <UL style={[tw`m-0 gap-2`, isSmallScreen && tw`mb-2`]}>
-            {data.libraries
-              .filter(({ npmPkg }) => npmPkg !== library.npmPkg)
-              .slice(0, LIMIT - 1)
-              .map(({ npmPkg, npm, github, unmaintained }) => (
-                <LI key={npmPkg}>
-                  <A
-                    href={`/package/${npmPkg}`}
-                    style={[
-                      tw`flex flex-row items-center justify-between rounded-lg border border-solid border-palette-gray2 px-4 py-2.5 no-underline dark:border-default`,
-                      isSmallScreen && tw`px-3 pt-2`,
-                    ]}
-                    hoverStyle={tw`bg-palette-gray1 dark:bg-dark`}>
-                    <View
-                      style={[
-                        tw`flex w-full max-w-full flex-row items-center gap-2`,
-                        isSmallScreen && tw`flex-col items-start gap-1`,
-                      ]}>
-                      <Caption style={tw`flex flex-shrink-0 items-center gap-1.5 text-sm`}>
-                        {unmaintained && (
-                          <Tooltip
-                            trigger={
-                              <View>
-                                <WarningIcon style={tw`text-warning-dark dark:text-warning`} />
-                              </View>
-                            }>
-                            Unmaintained
-                          </Tooltip>
-                        )}
-                        {npmPkg}
-                      </Caption>
-                      {github.description ? (
-                        <Label numberOfLines={1} style={tw`font-light text-secondary`}>
-                          {emoji.emojify(github.description)}
-                        </Label>
-                      ) : (
-                        <Label
-                          numberOfLines={1}
-                          style={tw`font-light text-tertiary dark:text-palette-gray5`}>
-                          The package does not have a description defined.
-                        </Label>
-                      )}
-                      <View
-                        style={[
-                          tw`ml-auto flex-row gap-4 text-sm font-light leading-[14px] text-icon`,
-                          isSmallScreen && tw`ml-0 mt-1`,
-                        ]}>
-                        <View style={tw`flex-row items-center gap-1 tabular-nums`}>
-                          <StarIcon style={tw`size-4 text-tertiary dark:text-palette-gray5`} />
-                          <span>{NUMBER_FORMATTER.format(github.stats.stars)}</span>
-                        </View>
-                        <View style={tw`flex-row items-center gap-1 tabular-nums`}>
-                          <DownloadIcon style={tw`text-tertiary dark:text-palette-gray5`} />
-                          <span>{bigNumberFormatter(npm?.downloads ?? 0)}</span>
-                        </View>
-                      </View>
-                    </View>
-                  </A>
-                </LI>
-              ))}
-          </UL>
-          {isSmallScreen && data.total > LIMIT && (
-            <A href={`/packages?owner=${encodeURI(owner)}`} style={tw`text-center`}>
-              <Caption style={tw`font-light`}>See all packages</Caption>
-            </A>
-          )}
-        </>
-      )}
-    </>
-  );
-}
