+++ package.json
@@ -48,15 +48,15 @@
     "react-dom": "19.2.8",
     "react-easy-linkify": "^1.0.8",
     "react-native": "0.86.2",
-    "react-native-safe-area-context": "^5.8.1",
+    "react-native-safe-area-context": "^5.9.0",
     "react-native-svg": "^15.15.5",
     "react-native-web": "^0.21.2",
-    "react-shiki": "0.11.0",
+    "react-shiki": "0.11.1",
     "rehype-raw": "^7.0.0",
     "remark-emoji": "^5.0.2",
     "remark-gfm": "^4.0.1",
     "semver": "^7.8.5",
-    "swr": "^2.5.0",
+    "swr": "^2.5.1",
     "tailwindcss": "^3.4.19",
     "twrnc": "^4.16.0",
     "use-debounce": "^10.1.1"
@@ -79,7 +79,7 @@
     "next-images": "^1.8.5",
     "oxfmt": "^0.63.0",
     "oxlint": "^1.78.0",
-    "oxlint-plugin-react-doctor": "^0.9.11",
+    "oxlint-plugin-react-doctor": "^0.9.12",
     "oxlint-tsgolint": "^7.0.2001",
     "simple-git-hooks": "^2.13.1",
     "typescript": "^7.0.2",
